requirements.txt
fastapi
uvicorn
pydantic
google-api-python-client
pytest
